update schema_dev, add AR 6, drop ruby < 2.5, add ruby 3.0
