+++ .gitignore
@@ -2,6 +2,7 @@
 /tmp
 /pkg
 /Gemfile.local
+/.idea
 
 *.lock
 *.log
