--- .travis.yml
@@ -1,89 +1,0 @@
-# This file was auto-generated by the schema_dev tool, based on the data in
-#                 ./schema_dev.yml
-# Please do not edit this file; any changes will be overwritten next time
-# schema_dev gets run.
----
-rvm:
-- 2.5.9
-- 2.7.3
-gemfile:
-- gemfiles/activerecord-5.2/Gemfile.sqlite3
-before_script: bundle exec rake create_databases
-after_script: bundle exec rake drop_databases
-script: bundle exec rake travis
-jobs:
-  include:
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.mysql2
-    rvm: 2.5.9
-    services:
-    - mysql
-    env: MYSQL_DB_USER=travis
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.5.9
-    addons:
-      postgresql: '10'
-      apt:
-        packages:
-        - postgresql-10
-        - postgresql-client-10
-    env: POSTGRESQL_DB_USER=postgres
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.5.9
-    addons:
-      postgresql: '11'
-      apt:
-        packages:
-        - postgresql-11
-        - postgresql-client-11
-    env: POSTGRESQL_DB_USER=travis PGPORT=5433
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.5.9
-    addons:
-      postgresql: '12'
-      apt:
-        packages:
-        - postgresql-12
-        - postgresql-client-12
-    env: POSTGRESQL_DB_USER=travis PGPORT=5433
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.5.9
-    addons:
-      postgresql: '9.6'
-    env: POSTGRESQL_DB_USER=postgres
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.mysql2
-    rvm: 2.7.3
-    services:
-    - mysql
-    env: MYSQL_DB_USER=travis
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.7.3
-    addons:
-      postgresql: '11'
-      apt:
-        packages:
-        - postgresql-11
-        - postgresql-client-11
-    env: POSTGRESQL_DB_USER=travis PGPORT=5433
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.7.3
-    addons:
-      postgresql: '12'
-      apt:
-        packages:
-        - postgresql-12
-        - postgresql-client-12
-    env: POSTGRESQL_DB_USER=travis PGPORT=5433
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.7.3
-    addons:
-      postgresql: '10'
-      apt:
-        packages:
-        - postgresql-10
-        - postgresql-client-10
-    env: POSTGRESQL_DB_USER=postgres
-  - gemfile: gemfiles/activerecord-5.2/Gemfile.postgresql
-    rvm: 2.7.3
-    addons:
-      postgresql: '9.6'
-    env: POSTGRESQL_DB_USER=postgres
