+++ Gemfile
@@ -2,4 +2,5 @@ source "http://rubygems.org"
 
 gemspec
 
-File.exist?(gemfile_local = File.expand_path('../Gemfile.local', __FILE__)) and eval File.read(gemfile_local), binding, gemfile_local
+gemfile_local = File.expand_path '../Gemfile.local', __FILE__
+eval File.read(gemfile_local), binding, gemfile_local if File.exist? gemfile_local
