+++ README.md
@@ -1,7 +1,6 @@
 [![Gem Version](https://badge.fury.io/rb/schema_plus_views.svg)](http://badge.fury.io/rb/schema_plus_views)
-[![Build Status](https://secure.travis-ci.org/SchemaPlus/schema_plus_views.svg)](http://travis-ci.org/SchemaPlus/schema_plus_views)
-[![Coverage Status](https://img.shields.io/coveralls/SchemaPlus/schema_plus_views.svg)](https://coveralls.io/r/SchemaPlus/schema_plus_views)
-[![Dependency Status](https://gemnasium.com/lomba/schema_plus_views.svg)](https://gemnasium.com/SchemaPlus/schema_plus_views)
+[![Build Status](https://github.com/SchemaPlus/schema_plus_views/actions/workflows/pr.yml/badge.svg)](http://github.com/SchemaPlus/schema_plus_views/actions)
+[![Coverage Status](https://coveralls.io/github/SchemaPlus/schema_plus_views/badge.svg)](https://coveralls.io/github/SchemaPlus/schema_plus_views)
 
 # SchemaPlus::Views
 
@@ -28,8 +27,11 @@ SchemaPlus::Views is tested on:
 
 <!-- SCHEMA_DEV: MATRIX - begin -->
 <!-- These lines are auto-generated by schema_dev based on schema_dev.yml -->
-* ruby **2.5.9** with activerecord **5.2**, using **mysql2**, **sqlite3** or **postgresql**
-* ruby **2.7.3** with activerecord **5.2**, using **mysql2**, **sqlite3** or **postgresql**
+* ruby **2.5** with activerecord **5.2**, using **mysql2**, **sqlite3**, **postgresql:9.6**, **postgresql:10**, **postgresql:11** or **postgresql:12**
+* ruby **2.5** with activerecord **6.0**, using **mysql2**, **sqlite3**, **postgresql:9.6**, **postgresql:10**, **postgresql:11** or **postgresql:12**
+* ruby **2.7** with activerecord **5.2**, using **mysql2**, **sqlite3**, **postgresql:9.6**, **postgresql:10**, **postgresql:11** or **postgresql:12**
+* ruby **2.7** with activerecord **6.0**, using **mysql2**, **sqlite3**, **postgresql:9.6**, **postgresql:10**, **postgresql:11** or **postgresql:12**
+* ruby **3.0** with activerecord **6.0**, using **mysql2**, **sqlite3**, **postgresql:9.6**, **postgresql:10**, **postgresql:11** or **postgresql:12**
 
 <!-- SCHEMA_DEV: MATRIX - end -->
 
@@ -212,7 +214,7 @@ Some things to know about to help you develop and test:
 * **schema_dev**:  SchemaPlus::Views uses [schema_dev](https://github.com/SchemaPlus/schema_dev) to
   facilitate running rspec tests on the matrix of ruby, activerecord, and database
   versions that the gem supports, both locally and on
-  [travis-ci](http://travis-ci.org/SchemaPlus/schema_plus_views)
+  [github actions](https://github.com/SchemaPlus/schema_plus_views/actions)
 
   To to run rspec locally on the full matrix, do:
 
@@ -223,7 +225,6 @@ Some things to know about to help you develop and test:
 
   The matrix of configurations is specified in `schema_dev.yml` in
   the project root.
-
 
 <!-- SCHEMA_DEV: TEMPLATE USES SCHEMA_DEV - end -->
